Finalizacion de migraciones

+++ urbanizacion/database/migrations/2021_06_05_221414_create_cuota_table.php
@@ -15,9 +15,9 @@ class CreateCuotaTable extends Migration
     {
         Schema::create('cuota', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger("id_credito");
+            $table->unsignedBigInteger("id_credito");
             $table->foreign('id_credito')->on('credito')->references('id')->onDelete('cascade');
-            $table->unsignedInteger("id_mora");
+            $table->unsignedBigInteger("id_mora");
             $table->foreign('id_mora')->on('mora')->references('id')->onDelete('cascade');
             $table->float("amortizacion");
             $table->integer("monto_cuota");
